FIX1/2: BdvViewUpdateBlenderSenderThread always sleeps a bit, "less busy" waiting:
Previously, if there were some pending updates, the construction of the code
prevented the main loop from sleeping, making it heavily checking if enough
time has passed since the last update to conduct the transfer. The new change
makes the loop always wait, even when there are pending updates to be transfered.