ADD2/2: BdvViewUpdateBlenderSenderThread monitors times passed since the last transfer:
The previous solution was effectively not transferring anything until sufficient
amount of time has passed between the last update request and now. If user
was "updating" the BDV (e.g. by panning the view in "a circle") for long time,
the required period of silence hasn't occured and no update was carried. This
wasn't great behaviour and so the code now monitors when it has sent anything
last time, and if "time enough" has passed, it sends one shot of the current state.
It doesn't interfere with the waiting for the period of silence.